Update program structure
